Moved RestoreAnalogTitleScroll to TitleMenu.cpp
